Moved reddit-parser into package util/reddit/
